Added code to set hostname and domainname

+++ goat.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 # Name:         goat (General OOB Automation Tool)
-# Version:      0.2.0
+# Version:      0.2.1
 # Release:      1
 # License:      CC-BA (Creative Commons By Attribution)
 #               http://creativecommons.org/licenses/by/4.0/legalcode
@@ -105,6 +105,7 @@ parser.add_argument("--avail",required=False)         # Get available version fr
 parser.add_argument("--check",required=False)         # Check current version against available version from vendor (e.g. BIOS)
 parser.add_argument("--model",required=False)         # Specify model (can be used with --avail)
 parser.add_argument("--port",required=False)          # Specify port to run service on
+parser.add_argument("--value",required=False)         # Set value
 parser.add_argument("--version",action='store_true')  # Display version
 parser.add_argument("--insecure",action='store_true') # Use HTTP/Telnet
 parser.add_argument("--verbose",action='store_true')  # Enable verbose output
@@ -374,7 +375,7 @@ def get_amt_value(get_value,ip,username,password,driver,http_proto,search):
 
 # Set AMT value
 
-def set_amt_value(set_value,ip,username,password,driver,http_proto,search):
+def set_amt_value(set_value,ip,username,password,driver,http_proto,search,value):
   if http_proto == "http":
     port_no = "16992"
   else:
@@ -382,22 +383,34 @@ def set_amt_value(set_value,ip,username,password,driver,http_proto,search):
   base_url = "%s://%s:%s@%s:%s" % (http_proto,username,password,ip,port_no)
   if re.search(r"power|reset",set_value):
     full_url = "%s/remote.htm" % (base_url)
+  if re.search(r"hostname|domainname",set_value):
+    full_url = "%s/fqdn.htm" % (base_url)
   driver.get(full_url)
-  if set_value == "poweroff":
-    driver.find_element_by_xpath('//input[@value="1"]').click()
-  if set_value == "powercycle":
-    driver.find_element_by_xpath('//input[@value="3"]').click()
-  if set_value == "reset":
-    driver.find_element_by_xpath('//input[@value="4"]').click()
   from selenium.webdriver.common.by import By
-  driver.find_element_by_xpath('//input[@value="Send Command"]').click()
-  time.sleep(2)
-  object = driver.switch_to.alert
-  time.sleep(2)
-  object.accept()
+  if re.search(r"power|reset",set_value):
+    if set_value == "poweroff":
+      driver.find_element_by_xpath('//input[@value="1"]').click()
+    if set_value == "powercycle":
+      driver.find_element_by_xpath('//input[@value="3"]').click()
+    if set_value == "reset":
+      driver.find_element_by_xpath('//input[@value="4"]').click()
+    driver.find_element_by_xpath('//input[@value="Send Command"]').click()
+    time.sleep(2)
+    object = driver.switch_to.alert
+    time.sleep(2)
+    object.accept()
+    string = "Sending %s to %s (Intel AMT has a 30s pause before operation is done)" % (set_value,ip)
+    print(string)
+  if re.search(r"hostname|domainname",set_value):
+    if re.search("hostname",set_value):
+      search = "HostName"
+    else:
+      search = "DomainName"
+    field = driver.find_element_by_name(search)
+    field.clear()
+    field.send_keys(value)
+    driver.find_element_by_xpath('//input[@value="   Submit   "]').click()
   driver.quit()
-  string = "Sending %s to %s (Intel AMT has a 30s pause before operation is done)" % (set_value,ip)
-  print(string)
   return
 
 # Compare versions
@@ -654,6 +667,11 @@ if option["port"]:
 if option["mesh"]:
   option["type"] = "amt"
 
+# Handle value switch
+
+if option["value"]:
+  value = option["value"]
+
 # Handle vendor switch
 
 if option["type"]:
@@ -698,7 +716,7 @@ if option["type"]:
       if option["get"]:
         get_amt_value(get_value,ip,username,password,driver,http_proto,search)
       if option["set"]:
-        set_amt_value(set_value,ip,username,password,driver,http_proto,search)
+        set_amt_value(set_value,ip,username,password,driver,http_proto,search,value)
 else:
   print("No OOB type specified")
   exit()
